refactor(package): 共有文字列の生成を main プロセスへ移設
apm.json のコアバージョン取得とインストール済みパッケージ一覧の整形を
services/packages.ts の buildShareString に集約し、tRPC
packages.getShareString query として公開する。renderer の sharePackages は
クリップボード書き込みとボタン遷移のみに縮小(ApmJson / programs /
shareStringVersion への依存を削除)。
